+++ ChangeLog.md
@@ -15,6 +15,7 @@
 * Remove Docker `pass-host` option
 * Don't require cabal-install to upload [#313](https://github.com/commercialhaskell/stack/issues/313)
 * Generate indexes for all deps and all installed snapshot packages [#143](https://github.com/fpco/commercialhaskell/issues/143)
+* Provide `--resolver global` option [#645](https://github.com/commercialhaskell/stack/issues/645)
 
 Bug fixes: